RHEL based f8-docs in staging
As part of an effort to migrate the services running in OSIO from CentOS
to RHEL, we are ready to push out the RHEL based image for this service
to staging.

In this commit we are keeping the prod environment on CentOS but
shifting
the staging environment to RHEL in order to validate.