Adds IMAGE parameters to several f8 services
This PR is part of an effort to migrate the services running in OSIO
from CentOS to RHEL.

This commit adds IMAGE as an environment parameter. At this stage it's
not currently being used by the service's openshift template, so this
commit does not affect staging or prod environments in any way.

However this enables the possibility of defining different urls for the
image in staging and prod, as soon as the service's openshift template
is updated.