Deploy che-tenant-maintainer into prod from Quay
Since this project has been deployed succesfully into staging from Quay, we can now promote to prod.

Note that the images are no longer being pused to the devshift registry, so if this PR is not merged, please make sure that in the next hash update you are also updating the image to be pulled from quay, instead of from the devshift registry.