Update core.yaml
commit https://github.com/fabric8-services/fabric8-wit/commit/2661cf80beff4aad9233b4a728b37b35e6298bcb
Author: Konrad Kleine <193408+kwk@users.noreply.github.com>
Date:   Tue Aug 7 15:40:55 2018 +0200

Provide join lock down  (fabric8-services/fabric8-wit#2211)
    
This includes an extra-condition in the `ON` part of the table `JOINS` for areas, codebases and iterations to only join those tables filtered by their space ID. I'm not sure though if this really fixes the problem (see https://github.com/fabric8-services/fabric8-wit/issues/2210#issuecomment-410241026).
    
## TODO
    
As of yesterday's (07.08.2018) discussion with @aslakknutsen we did experiments and found that in order to keep the rows in the search small, we have to establish a condition on the final SQL `WHERE` clause that limits the selection to work items from a particular space. At the moment, the current `/api/search` endpoint is so generic that it doesn't require a limitation by space on the root of the `WHERE` clause. That's why @aslakknutsen and I agreed to create a search endpoint under `/api/spaces/<SPACE-UUID>/search` in order to automatically add the space ID to the query condition. This will be implemented in another PR and is tracked in https://github.com/openshiftio/openshift.io/issues/4124
    
See fabric8-services/fabric8-wit#2210.